+++ packages/espressocash_app/lib/features/incoming_split_key_payments/routes.dart
@@ -1,6 +1,6 @@
 import 'package:auto_route/auto_route.dart';
 
-import 'src/widgets/first_part_qr.dart';
+import 'src/widgets/first_part_qr_screen.dart';
 import 'src/widgets/first_part_ready_screen.dart';
 import 'src/widgets/incoming_split_key_payment_screen.dart';
